configure webpack and babel
the tooling now includes yarn itself (which can be relied on, since netlify has yarn).

closes #330

The structure of the JS files now is (in my opinion) more reasonable to think about, and also will allow for faster development for further features.

jQuery, bootstrap are still loaded via cdn, because that's better for large bundles like those.

instantsearch, docsearch, moment are loaded via npm and bundled in (which saved approx. 300kb on the search page)